requirements.txt
grpcio-tools==1.73.1
pydantic==2.11.7
grpcio-reflection==1.73.1